build: #65 create generator for redux minus test file
