+++ package.json
@@ -13,7 +13,8 @@
     "postinstall": "husky install",
     "create:component": "turbo gen component",
     "create:package": "turbo gen package",
-    "create:app": "turbo gen app"
+    "create:app": "turbo gen app",
+    "create:redux": "turbo gen redux"
   },
   "devDependencies": {
     "@turbo/gen": "^1.10.2",
