+++ packages/shared/core/src/index.tsx
@@ -1,2 +1,3 @@
+/* PLOP_INJECT_EXPORT */
 export * from './stores/theme';
 export * from './utils/local-authorization/local-authorization';
