+++ turbo/generators/templates/store/store.hbs
@@ -1,0 +1,11 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+export interface {{pascalCase name}}Interface {}
+
+const initialState: {{pascalCase name}}Interface = {};
+
+export const {{camelCase name}}Slice = createSlice({
+  name: '{{camelCase name}}',
+  initialState,
+  reducers: {},
+});
